# [Misc] add structure logging, write to file and log tracing for SGL R…