--- a/sgl-router/py_test/test_launch_router.py
+++ b/sgl-router/py_test/test_launch_router.py
@@ -37,6 +37,7 @@ class TestLaunchRouter(unittest.TestCase):
             max_tree_size=2**24,
             max_payload_size=4 * 1024 * 1024,  # 4MB
             verbose=False,
+            log_dir=None,
         )
 
     def create_router_args(self, **kwargs):