<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->
This PR implements logging for the SGL Router using the tracing. The logging system provides structured logs with timestamps, context information, and configurable log levels.
- Structured logging with multiple log levels (TRACE, DEBUG, INFO, WARN, ERROR)
- Simultaneous logging to console and files
- JSON output format option for log analysis tools
- Minimal overhead implementation with non-blocking I/O
Safe initialization that prevents crashes when initialized multiple times

## Modifications

<!-- Describe the changes made in this PR. -->
- Uses `tracing`, `tracing-subscriber`, `tracing-log`, and `tracing-appender` crates
- Updated documentation in `README.md`
- Added new `logging.rs`

## Performance Benchmark
To ensure the logging implementation has minimal overhead, extensive benchmarking was performed. The benchmark measured the overhead of logging operations with different configurations:

- Created a benchmark that executes the same operation with and without logging enabled
- Measured the time per operation for:
  - No logging
  - Console-only logging
  - File logging (daily rolling files)

Results
```
Without logging (total): 99ns for 1,000,000 ops
With console logging (total): 395.498µs for 100,000 ops
With file logging (total): 2.018893ms for 100,000 ops

Time per operation without logging: 0.00 ns
Time per operation with console logging: 3.95 ns
Time per operation with file logging: 20.19 ns
```

- Console logging: Adds only ~4 nanoseconds per log operation
- File logging: Adds ~20 nanoseconds per log operation
- These overheads are negligible for most operations and will not impact the router's performance

![Screenshot 2025-04-24 at 11 51 49 PM](https://github.com/user-attachments/assets/fd90a573-bf0b-4655-8755-dd562850d52b)


## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [x] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [x] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [x] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [x] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.