@slin1237 2025-04-25
havent added all flags to the entry point yet, most of them are hardcoded in the code at the moment
we can turn them on later if we have to, such as json logging etc